editing is great, but now posting acting up
