+++ src/api/entries.json
@@ -1,11 +1,11 @@
 {
   "entries": [
     {
-      "id": 1,
       "date": "2020-05-29",
       "concepts": "New beginnings",
-      "entry": "Only three weeks in and NSS has already been an immensly rewarding experience.",
-      "mood": "😀"
+      "entry": "Only three weeks in and NSS has already been an immensly rewarding experience.\n\n\nwoooo i'm from the future !!! sppooooky 👻",
+      "mood": "😀",
+      "id": 1
     },
     {
       "date": "2020-06-19",
@@ -17,7 +17,7 @@
     {
       "date": "2020-06-24",
       "concepts": "forking ",
-      "entry": "Forks: great for many things, but terrible for soup (except ramen)",
+      "entry": "Forks: great for many things, but terrible for soup (except ramen)\n\nedit: \n\nAlso chili, if it's thick.",
       "mood": "🎯",
       "id": 3
     },
