+++ src/scripts/clearForm.js
@@ -1,0 +1,8 @@
+export default () => {
+                //clearing out the form
+                document.querySelector("#entryId").value = ""
+                document.querySelector("#journalDate").value = ""
+                document.querySelector("#journalConcepts").value = ""
+                document.querySelector("#journalEntry").value = ""
+                document.querySelector("#mood").value = ""
+}
