+++ src/scripts/updateEntry.js
@@ -1,4 +1,5 @@
-import API from "./data.js"
+import API from "./data.js" 
+import entryGrabber from "./entryGrabber.js"
 
 const updateFormFields = entryId => {
